cleaning up code with Flake8

+++ testapi/routes.py
@@ -7,10 +7,7 @@ from testapi.models import Widget
 from datetime import datetime
 
 
-
-
 # create databse tabels if they're not already there
-
 from testapi import db
 try:
     db.create_all()
@@ -19,7 +16,7 @@ except:
     print("Table creation failed.")
 
 
-#helper function to strip db rows of the "row" object.
+# helper function to strip db rows of the "row" object.
 def rowToDict(row):
     x = {}
     for col in row.__table__.columns:
@@ -28,24 +25,20 @@ def rowToDict(row):
     return x
 
 
-
-#default route. nothing to see here as this is an api and does not contain front end work. 
+# default route. nothing to see here as this is an api and does not contain front end work.
 @app.route("/", methods=["GET", "POST"])
 def hello():
     return "hello all!"
 
-# to retreive records for all widgets in the DB. 
+
+# to retreive records for all widgets in the DB.
 @app.route("/widget", methods=["GET"])
 @app.route("/widget/", methods=["GET"])
 def widget():
     result = []
-    widgs = Widget.query.with_entities(Widget.id, Widget.name, Widget.parts, Widget.date_create, Widget.date_update).all()
-    #widgs = Widget.query.first()
     for w in Widget.query.all():
         result.append(rowToDict(w))
     return result
-
-    
 
 
 # To retrieve a record for a single widget given a name
@@ -53,12 +46,16 @@ def widget():
 @app.route("/widget/<wname>/", methods=['GET', 'DELETE'])
 def get_widget(wname):
     w = Widget.query.filter_by(name=wname).first()
-    if request.method == "GET":   
+
+    if (w is None):
+        return ("<p>The specified entry does not exist.<p>")
+    if request.method == "GET":
         return w.as_dict()
     if request.method == "DELETE":
         db.session.delete(w)
         db.session.commit()
-        return 
+        return "<p> Record Deleted <p>"
+
 
 # to create or update a record for a single widget in the DB
 @app.route("/widget/<wname>/<int:no_parts>", methods=['POST'])
@@ -67,22 +64,13 @@ def alter_Widget(wname, no_parts):
     w = Widget.query.filter_by(name=wname).first()
     print(type(w))
 
-    if(w == None): #if the record doesn't exist, let's create it.
-        widg = Widget(name = wname, parts = no_parts)
+    if (w is None):  # if the record doesn't exist, let's create it.
+        widg = Widget(name=wname, parts=no_parts)
         db.session.add(widg)
         db.session.commit()
         return "<p>Record created</p>"
-    elif(type(w) != None):                  #if the record already exists...update the existing one.
+    elif (type(w) is not None):  # if the record already exists...update the existing one.
         w.parts = no_parts
         w.date_update = datetime.utcnow()
         db.session.commit()
         return "<p>Record Updated.</p>"
-
-        
-        
-        
-    
-
-
-
-
